renames and whatnot

+++ LoginPageTests/AuthServiceTests.cs
@@ -1,19 +1,19 @@
 using Hospital.Configs;
-using Hospital.DatabaseServices;
-using Hospital.Managers;
 using Hospital.Models;
+using Hospital.Repositories;
+using Hospital.Services;
 using Microsoft.Data.SqlClient;
 
 namespace LoginPageTests;
 
 [TestClass]
-public class AuthManagerModelTests
+public class AuthServiceTests
 {
-    IAuthManagerModel authManagerModel;
-    public AuthManagerModelTests()
+    IAuthService _authService;
+    public AuthServiceTests()
     {
-        ILogInDatabaseService logInDatabaseService = new LogInDatabaseService();
-        authManagerModel = new AuthManagerModel(logInDatabaseService);
+        ILogInRepository logInRepository = new LogInRepository();
+        _authService = new AuthService(logInRepository);
     }
 
     [TestMethod]
@@ -33,10 +33,10 @@ public class AuthManagerModelTests
                 180
             );
 
-        bool result = await authManagerModel.CreateAccount(model);
+        bool result = await _authService.CreateAccount(model);
         Assert.IsTrue(result);
 
-        string query = "DELETE FROM Users WHERE Username = @username";
+        string query = "DELETE FROM Logs WHERE UserId=(SELECT TOP 1 UserId FROM Users WHERE Username=@username)";
 
         using SqlConnection connectionToDatabase = new SqlConnection(Config.GetInstance().DatabaseConnection);
         await connectionToDatabase.OpenAsync().ConfigureAwait(false);
@@ -44,7 +44,12 @@ public class AuthManagerModelTests
         using SqlCommand command = new SqlCommand(query, connectionToDatabase);
         command.Parameters.AddWithValue("@username", "nexuser");
 
+        query = "DELETE FROM Users WHERE Username=@username";
         int rowsAffected = await command.ExecuteNonQueryAsync().ConfigureAwait(false);
+
+        using SqlCommand command2 = new SqlCommand(query, connectionToDatabase);
+        command2.Parameters.AddWithValue("@username", "nexuser");
+        int rowsAffected2 = await command2.ExecuteNonQueryAsync().ConfigureAwait(false);
     }
 
     [TestMethod]
@@ -66,7 +71,7 @@ public class AuthManagerModelTests
 
         await Assert.ThrowsExceptionAsync<Hospital.Exceptions.AuthenticationException>(async () =>
         {
-            await authManagerModel.CreateAccount(model);
+            await _authService.CreateAccount(model);
         });
     }
 
@@ -89,7 +94,7 @@ public class AuthManagerModelTests
 
         await Assert.ThrowsExceptionAsync<Hospital.Exceptions.AuthenticationException>(async () =>
         {
-            await authManagerModel.CreateAccount(model);
+            await _authService.CreateAccount(model);
         });
     }
 
@@ -112,7 +117,7 @@ public class AuthManagerModelTests
 
         await Assert.ThrowsExceptionAsync<Hospital.Exceptions.AuthenticationException>(async () =>
         {
-            await authManagerModel.CreateAccount(model);
+            await _authService.CreateAccount(model);
         });
     }
 
@@ -135,7 +140,7 @@ public class AuthManagerModelTests
 
         await Assert.ThrowsExceptionAsync<Hospital.Exceptions.AuthenticationException>(async () =>
         {
-            await authManagerModel.CreateAccount(model);
+            await _authService.CreateAccount(model);
         });
     }
 
@@ -143,7 +148,7 @@ public class AuthManagerModelTests
     public async Task TestLoadUserbyUsername_WithValidUsername_ReturnsTrue()
     {
         // Task<bool> LoadUserByUsername(string username);
-        var result = await authManagerModel.LoadUserByUsername("john_doe");
+        var result = await _authService.LoadUserByUsername("john_doe");
         Assert.IsTrue(result);
     }
 
@@ -153,7 +158,7 @@ public class AuthManagerModelTests
         // Task<bool> LoadUserByUsername(string username);
         await Assert.ThrowsExceptionAsync<Hospital.Exceptions.AuthenticationException>(async () =>
         {
-            await authManagerModel.LoadUserByUsername("not_john_doe");
+            await _authService.LoadUserByUsername("not_john_doe");
         });
     }
 
@@ -161,7 +166,7 @@ public class AuthManagerModelTests
     public async Task VerifyPassword_IncorrectPassword_ReturnsFalse()
     {
         // Task<bool> VerifyPassword(string userInputPassword);
-        var result = await authManagerModel.VerifyPassword("wrong_password");
+        var result = await _authService.VerifyPassword("wrong_password");
         Assert.IsFalse(result);
     }
